powerpoint-tools-mcp/requirements.txt
fastmcp
uvicorn
python-pptx
